Refactored topology tests and property tests into archipelago tests
