+++ PyGMO/test/CMakeLists.txt
@@ -1,4 +1,4 @@
 INSTALL(FILES __init__.py DESTINATION ${PYGMO_INSTALL_PATH}/test/)
 INSTALL(FILES _hypervolume_tests.py DESTINATION ${PYGMO_INSTALL_PATH}/test/)
 INSTALL(FILES _topology_tests.py DESTINATION ${PYGMO_INSTALL_PATH}/test/)
-INSTALL(FILES _property_tests.py DESTINATION ${PYGMO_INSTALL_PATH}/test/)
+INSTALL(FILES _archipelago_tests.py DESTINATION ${PYGMO_INSTALL_PATH}/test/)
