+++ PyGMO/test/__init__.py
@@ -117,13 +117,13 @@ def run_full_test_suite():
     from PyGMO import test
     from PyGMO.test._hypervolume_tests import get_hv_suite
     from PyGMO.test._topology_tests import get_topology_test_suite
-    from PyGMO.test._property_tests import get_property_test_suite
+    from PyGMO.test._archipelago_tests import get_archipelago_test_suite
     suite = _ut.TestLoader().loadTestsFromModule(test)
 
     # Add external suites explicitly
     suite.addTests(get_hv_suite())
     suite.addTests(get_topology_test_suite())
-    suite.addTests(get_property_test_suite())
+    suite.addTests(get_archipelago_test_suite())
 
     _ut.TextTestRunner(verbosity=2).run(suite)
 
@@ -135,6 +135,11 @@ def run_hv_test_suite():
 
 
 def run_topology_test_suite():
-    """Run the hypervolume test suite."""
+    """Run the topology test suite."""
     from PyGMO.test._topology_tests import get_topology_test_suite
     _ut.TextTestRunner(verbosity=2).run(get_topology_test_suite())
+
+def run_archipelago_test_suite():
+    """Run the archipelago test suite."""
+    from PyGMO.test._archipelago_tests import get_archipelago_test_suite
+    _ut.TextTestRunner(verbosity=2).run(get_archipelago_test_suite())
